+++ TabletopTweaks-Base/Bugfixes/Items/Cooking.cs
@@ -1,0 +1,62 @@
+﻿using HarmonyLib;
+using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Spells;
+using Kingmaker.Blueprints.Items.Ecnchantments;
+using Kingmaker.Blueprints.Items.Equipment;
+using Kingmaker.Blueprints.JsonSystem;
+using Kingmaker.Blueprints.Loot;
+using Kingmaker.Designers.EventConditionActionSystem.Actions;
+using Kingmaker.Designers.Mechanics.Facts;
+using Kingmaker.ElementsSystem;
+using Kingmaker.EntitySystem.Entities;
+using Kingmaker.Enums;
+using Kingmaker.Enums.Damage;
+using Kingmaker.Items;
+using Kingmaker.RuleSystem;
+using Kingmaker.RuleSystem.Rules.Damage;
+using Kingmaker.UnitLogic.Abilities.Blueprints;
+using Kingmaker.UnitLogic.Abilities.Components.AreaEffects;
+using Kingmaker.UnitLogic.ActivatableAbilities;
+using Kingmaker.UnitLogic.Buffs.Blueprints;
+using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Mechanics;
+using Kingmaker.UnitLogic.Mechanics.Actions;
+using Kingmaker.UnitLogic.Mechanics.Components;
+using Kingmaker.UnitLogic.Mechanics.Conditions;
+using Kingmaker.Utility;
+using System.Linq;
+using TabletopTweaks.Core.NewActions;
+using TabletopTweaks.Core.NewComponents;
+using TabletopTweaks.Core.NewComponents.OwlcatReplacements;
+using TabletopTweaks.Core.Utilities;
+using static TabletopTweaks.Base.Main;
+
+namespace TabletopTweaks.Base.Bugfixes.Items {
+    internal class Cooking {
+        [HarmonyPatch(typeof(BlueprintsCache), "Init")]
+        static class BlueprintsCache_Init_Patch {
+            static bool Initialized;
+
+            static void Postfix() {
+                if (Initialized) return;
+                Initialized = true;
+                TTTContext.Logger.LogHeader("Patching Cooking");
+                PatchScreamingOmelet();
+            }
+            static void PatchScreamingOmelet() {
+                if (TTTContext.Fixes.Items.Cooking.IsDisabled("ScreamingOmelet")) { return; }
+
+                var ScreamingOmeletBuff = BlueprintTools.GetBlueprint<BlueprintBuff>("0c283d43dde431f43aa13ace4488fba4");
+
+                ScreamingOmeletBuff.TemporaryContext(bp => {
+                    bp.GetComponent<AddContextStatBonus>()?.TemporaryContext(c => {
+                        c.Multiplier = 2;
+                    });
+                });
+
+                TTTContext.Logger.LogPatch(ScreamingOmeletBuff);
+            }
+        }
+    }
+}